+++ packages/acmekit/src/modules/customer.ts
@@ -1,6 +1,0 @@
-import CustomerModule from "@acmekit/customer"
-
-export * from "@acmekit/customer"
-
-export default CustomerModule
-export const discoveryPath = require.resolve("@acmekit/customer")