+++ packages/acmekit/src/modules/store.ts
@@ -1,6 +1,0 @@
-import StoreModule from "@acmekit/store"
-
-export * from "@acmekit/store"
-
-export default StoreModule
-export const discoveryPath = require.resolve("@acmekit/store")