+++ packages/admin/dashboard/src/hooks/api/api-keys.tsx
@@ -11,7 +11,6 @@ import {
 import { sdk } from "../../lib/client"
 import { queryClient } from "../../lib/query-client"
 import { queryKeysFactory } from "../../lib/query-key-factory"
-import { salesChannelsQueryKeys } from "./sales-channels"
 
 const API_KEYS_QUERY_KEY = "api_keys" as const
 export const apiKeysQueryKeys = queryKeysFactory(API_KEYS_QUERY_KEY)
@@ -127,53 +126,5 @@ export const useDeleteApiKey = (
 
       options?.onSuccess?.(data, variables, context)
     },
-  })
-}
-
-export const useBatchRemoveSalesChannelsFromApiKey = (
-  id: string,
-  options?: UseMutationOptions<
-    HttpTypes.AdminApiKeyResponse,
-    FetchError,
-    HttpTypes.AdminBatchLink["remove"]
-  >
-) => {
-  return useMutation({
-    mutationFn: (payload) =>
-      sdk.admin.apiKey.batchSalesChannels(id, { remove: payload }),
-    onSuccess: (data, variables, context) => {
-      queryClient.invalidateQueries({ queryKey: apiKeysQueryKeys.lists() })
-      queryClient.invalidateQueries({ queryKey: apiKeysQueryKeys.detail(id) })
-      queryClient.invalidateQueries({
-        queryKey: salesChannelsQueryKeys.lists(),
-      })
-
-      options?.onSuccess?.(data, variables, context)
-    },
-    ...options,
-  })
-}
-
-export const useBatchAddSalesChannelsToApiKey = (
-  id: string,
-  options?: UseMutationOptions<
-    HttpTypes.AdminApiKeyResponse,
-    FetchError,
-    HttpTypes.AdminBatchLink["add"]
-  >
-) => {
-  return useMutation({
-    mutationFn: (payload) =>
-      sdk.admin.apiKey.batchSalesChannels(id, { add: payload }),
-    onSuccess: (data, variables, context) => {
-      queryClient.invalidateQueries({ queryKey: apiKeysQueryKeys.lists() })
-      queryClient.invalidateQueries({ queryKey: apiKeysQueryKeys.detail(id) })
-      queryClient.invalidateQueries({
-        queryKey: salesChannelsQueryKeys.lists(),
-      })
-
-      options?.onSuccess?.(data, variables, context)
-    },
-    ...options,
   })
 }