+++ packages/admin/dashboard/src/vite-env.d.ts
@@ -2,6 +2,8 @@
 
 interface ImportMetaEnv {
   readonly VITE_ACMEKIT_ADMIN_BACKEND_URL: string
+  readonly VITE_ACMEKIT_FRONTEND_URL: string
+  /** @deprecated Use VITE_ACMEKIT_FRONTEND_URL */
   readonly VITE_ACMEKIT_STOREFRONT_URL: string
   readonly VITE_ACMEKIT_V2: "true" | "false"
 }
@@ -14,7 +16,7 @@ interface ImportMeta {
 }
 
 declare const __BACKEND_URL__: string | undefined
-declare const __STOREFRONT_URL__: string | undefined
+declare const __FRONTEND_URL__: string | undefined
 declare const __BASE__: string
 declare const __AUTH_TYPE__: "session" | "jwt" | undefined
 declare const __JWT_TOKEN_STORAGE_KEY__: string | undefined