+++ packages/core/utils/src/common/__tests__/get-resolved-plugins.spec.ts
@@ -68,10 +68,6 @@ describe("getResolvedPlugins | relative paths", () => {
   })
 
   test("scan plugin modules", async () => {
-    await fs.createJson("node_modules/@acmekit/draft-order/package.json", {
-      name: "@acmekit/draft-order",
-      version: "1.0.0",
-    })
     await fs.createJson("plugins/dummy/package.json", {
       name: "my-dummy-plugin",
       version: "1.0.0",
@@ -96,35 +92,26 @@ describe("getResolvedPlugins | relative paths", () => {
       false
     )
 
-    expect(plugins).toEqual(expect.arrayContaining([
-      {
-        resolve: path.join(fs.basePath, "./plugins/dummy/.acmekit/server/src"),
-        admin: undefined,
-        name: "my-dummy-plugin",
-        id: "my-dummy-plugin",
-        options: { apiKey: "asecret" },
-        version: "1.0.0",
-        modules: [
-          {
-            options: {
-              apiKey: "asecret",
+    expect(plugins).toEqual(
+      expect.arrayContaining([
+        {
+          resolve: path.join(fs.basePath, "./plugins/dummy/.acmekit/server/src"),
+          admin: undefined,
+          name: "my-dummy-plugin",
+          id: "my-dummy-plugin",
+          options: { apiKey: "asecret" },
+          version: "1.0.0",
+          modules: [
+            {
+              options: {
+                apiKey: "asecret",
+              },
+              resolve: "./plugins/dummy/.acmekit/server/src/modules/blog",
             },
-            resolve: "./plugins/dummy/.acmekit/server/src/modules/blog",
-          },
-        ],
-      },
-      {
-        id: "@acmekit/draft-order",
-        modules: [],
-        name: "@acmekit/draft-order",
-        options: {},
-        resolve: path.join(
-          fs.basePath,
-          "node_modules/@acmekit/draft-order/.acmekit/server/src"
-        ),
-        version: "1.0.0",
-      },
-    ]))
+          ],
+        },
+      ])
+    )
   })
 
   test("throw error when package.json file is missing", async () => {
@@ -255,38 +242,25 @@ describe("getResolvedPlugins | package reference", () => {
       false
     )
 
-    expect(plugins).toEqual(expect.arrayContaining([
-      {
-        resolve: path.join(
-          fs.basePath,
-          "node_modules/@plugins/dummy/.acmekit/server/src"
-        ),
-        admin: undefined,
-        name: "my-dummy-plugin",
-        id: "my-dummy-plugin",
-        options: { apiKey: "asecret" },
-        version: "1.0.0",
-        modules: [],
-      },
-      {
-        id: "@acmekit/draft-order",
-        modules: [],
-        name: "@acmekit/draft-order",
-        options: {},
-        resolve: path.join(
-          fs.basePath,
-          "node_modules/@acmekit/draft-order/.acmekit/server/src"
-        ),
-        version: "1.0.0",
-      },
-    ]))
+    expect(plugins).toEqual(
+      expect.arrayContaining([
+        {
+          resolve: path.join(
+            fs.basePath,
+            "node_modules/@plugins/dummy/.acmekit/server/src"
+          ),
+          admin: undefined,
+          name: "my-dummy-plugin",
+          id: "my-dummy-plugin",
+          options: { apiKey: "asecret" },
+          version: "1.0.0",
+          modules: [],
+        },
+      ])
+    )
   })
 
   test("scan plugin modules", async () => {
-    await fs.createJson("node_modules/@acmekit/draft-order/package.json", {
-      name: "@acmekit/draft-order",
-      version: "1.0.0",
-    })
     await fs.createJson("package.json", {})
     await fs.createJson("node_modules/@plugins/dummy/package.json", {
       name: "my-dummy-plugin",
@@ -312,38 +286,29 @@ describe("getResolvedPlugins | package reference", () => {
       false
     )
 
-    expect(plugins).toEqual(expect.arrayContaining([
-      {
-        resolve: path.join(
-          fs.basePath,
-          "node_modules/@plugins/dummy/.acmekit/server/src"
-        ),
-        admin: undefined,
-        name: "my-dummy-plugin",
-        id: "my-dummy-plugin",
-        options: { apiKey: "asecret" },
-        version: "1.0.0",
-        modules: [
-          {
-            options: {
-              apiKey: "asecret",
+    expect(plugins).toEqual(
+      expect.arrayContaining([
+        {
+          resolve: path.join(
+            fs.basePath,
+            "node_modules/@plugins/dummy/.acmekit/server/src"
+          ),
+          admin: undefined,
+          name: "my-dummy-plugin",
+          id: "my-dummy-plugin",
+          options: { apiKey: "asecret" },
+          version: "1.0.0",
+          modules: [
+            {
+              options: {
+                apiKey: "asecret",
+              },
+              resolve: "@plugins/dummy/.acmekit/server/src/modules/blog",
             },
-            resolve: "@plugins/dummy/.acmekit/server/src/modules/blog",
-          },
-        ],
-      },
-      {
-        id: "@acmekit/draft-order",
-        modules: [],
-        name: "@acmekit/draft-order",
-        options: {},
-        resolve: path.join(
-          fs.basePath,
-          "node_modules/@acmekit/draft-order/.acmekit/server/src"
-        ),
-        version: "1.0.0",
-      },
-    ]))
+          ],
+        },
+      ])
+    )
   })
 
   test("throw error when package.json file is missing", async () => {