+++ packages/admin/admin-bundler/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2.13.9
+
+### Patch Changes
+
+- changes
+
+- Updated dependencies []:
+  - @acmekit/admin-shared@2.13.9
+  - @acmekit/admin-vite-plugin@2.13.9
+  - @acmekit/dashboard@2.13.9
+
 ## 2.13.8
 
 ### Patch Changes