+++ packages/admin/admin-vite-plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.13.9
+
+### Patch Changes
+
+- changes
+
+- Updated dependencies []:
+  - @acmekit/admin-shared@2.13.9
+
 ## 2.13.8
 
 ### Patch Changes