+++ packages/admin/dashboard/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.13.9
+
+### Patch Changes
+
+- changes
+
+- Updated dependencies []:
+  - @acmekit/admin-shared@2.13.9
+  - @acmekit/js-sdk@2.13.9
+  - @acmekit/icons@2.13.9
+  - @acmekit/ui@4.1.9
+
 ## 2.13.8
 
 ### Patch Changes