+++ packages/cli/oas/acmekit-oas-cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.13.9
+
+### Patch Changes
+
+- changes
+
+- Updated dependencies []:
+  - @acmekit/acmekit@2.13.9
+  - @acmekit/utils@2.13.9
+
 ## 2.13.8
 
 ### Patch Changes