+++ packages/core/modules-sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2.13.9
+
+### Patch Changes
+
+- changes
+
+- Updated dependencies []:
+  - @acmekit/orchestration@2.13.9
+  - @acmekit/utils@2.13.9
+  - @acmekit/deps@2.13.9
+
 ## 2.13.8
 
 ### Patch Changes