+++ packages/design-system/ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 4.1.9
+
+### Patch Changes
+
+- changes
+
+- Updated dependencies []:
+  - @acmekit/icons@2.13.9
+
 ## 4.1.8
 
 ### Patch Changes